--- server/db/assessments.js
@@ -1,21 +1,0 @@
-const environment = 'development'
-const config = require('../../knexfile')[environment]
-const connection = require('knex')(config)
-
-function getAssessments(id, db = connection) {}
-
-function updateEvidence(id, evidence, db = connection) {
-  return db('student_assessments')
-    .where('assessment_id', id)
-    .andWhere('student_id', evidence.studentId)
-    .update({
-      evidence: evidence.evidence,
-      date_modified: evidence.date
-    })
-    .catch(err => err)
-}
-
-module.exports = {
-  getAssessments,
-  updateEvidence
-}
