+++ server/db/assessmentsDb.js
@@ -12,6 +12,19 @@ function getAssessmentsById(id, db=connection){
     .where('assessments.id',id)
     .select()
 }
+
+function updateEvidence(id, evidence, db = connection) {
+    return db('student_assessments')
+      .where('assessment_id', id)
+      .andWhere('student_id', evidence.studentId)
+      .update({
+        evidence: evidence.evidence,
+        date_modified: evidence.date
+      })
+      .catch(err => err)
+  }
+
+
 module.exports ={
     getAssessmentsById,
     getAssessments,
